Ported to QT6/KF6. Interim.  Still more Player work.  Mute button works again.  Volume works properly again.

+++ src/gui/widgets/PlayerControls.h
@@ -44,7 +44,7 @@ public:
 
     QMediaPlayer::PlaybackState state() const;
 	bool isMuted() const { return m_playerMuted; }
-    int volume() const;
+    float volume() const;
 
 Q_SIGNALS:
     void play();
@@ -54,12 +54,12 @@ Q_SIGNALS:
     void previous();
 	void changeShuffle(bool);
 	void changeRepeat(bool);
-	void changeVolume(int);
+	void changeVolume(float);
     void changeMuting(bool);
 
 public Q_SLOTS:
     void setPlaybackState(QMediaPlayer::PlaybackState state);
-    void setVolume(int volume);
+    void setVolume(float volume);
     void setMuted(bool muted);
 	void onDurationChanged(qint64 duration);
 	void onPositionChanged(qint64 pos);
